pyproject.toml
[project]
name = "closure-qqbot"
version = "0.1.0"
description = "可露希尔工作室"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Criogaid", email = "kripto.sakura@gmail.com"},
]
dependencies = [
    "nonebot2[fastapi]>=2.3.3",
    "nonebot-plugin-apscheduler>=0.5.0",
    "pip>=24.2",
    "pipx>=1.7.1",
    "nonebot-adapter-onebot>=2.4.5",
    "nonebot-adapter-console>=0.6.0",
    "requests>=2.32.3",
    "cachetools>=5.5.0",
    "peewee>=3.17.6",
    "httpx>=0.27.2",
    "aiocache>=0.12.3",
    "nonebot-plugin-htmlrender>=0.3.5",
]
license = {text = "None"}

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
    {name = "Console", module_name = "nonebot.adapters.console"},
]
plugins = ["nonebot_plugin_apscheduler", "nonebot_plugin_htmlrender"]
plugin_dirs = ["plugins"]
builtin_plugins = ["echo"]

[tool.pdm]
distribution = true

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
